Move state out of manager
