+++ lib/ruby_jard.rb
@@ -15,6 +15,7 @@ require 'ruby_jard/key_binding'
 require 'ruby_jard/key_bindings'
 require 'ruby_jard/repl_processor'
 require 'ruby_jard/repl_manager'
+require 'ruby_jard/repl_state'
 require 'ruby_jard/screen_manager'
 require 'ruby_jard/reflection'
